refactoring mipstests to support both executor and prover tests
