+++ crates/core/executor/Cargo.toml
@@ -52,6 +52,7 @@ num_enum = "0.7.5"
 
 [dev-dependencies]
 test-artifacts = { path = "../../test-artifacts" }
+zkm-instruction-test-defs = { workspace = true }
 
 [features]
 bigint-rug = ["zkm-curves/bigint-rug"]
